Rename task directory to scheduler
Scheduler jobs could be based on scheduler task or extbase command
controller (maybe other too)
